Fixed floatLax in HtmlTable.Sort
Major bug here, the text was never getting converted into a float and
therefore could never get properly sorted.

+++ Source/Interface/HtmlTable.Sort.js
@@ -318,7 +318,7 @@ HtmlTable.Parsers = {
 	'floatLax': {
 		match: /^[^\d]+[\d]+\.[\d]+$/,
 		convert: function(){
-			return this.get('text').replace(/[^-?^\d.]/, '').stripTags();
+			return this.get('text').replace(/[^-?^\d.]/, '').stripTags().toFloat();
 		},
 		number: true
 	},
